youtube_autoplaylist/requirements.txt
httplib2==0.20.4
oauth2client==4.1.3

python-dateutil~=2.8.2
google-api-python-client~=2.46.0